src/index: Put _write in try/catch
This avoids the bad edge case of the logger throwing
an exception and the uncaught exception listener logging
the exception and then the logger throwing a new uncaught
exception which leads to an infinite loop.

We break the loop by passing the error to `this.onError`
and not to the uncaught exception.